// client/src/data/mining-lasers.ts
export interface MiningLaser {
  name: string;
  size: 0 | 1 | 2;
  price: number;
  optimumRange: number;
  maxRange: number;
  minPower: number;
  minPowerPct: number;
  maxPower: number;
  extractPower: number;
  moduleSlots: number;
  resistance: number;
  instability: number;
  optimalChargeRate: number;
  optimalChargeWindow: number;
  inertMaterials: number;
  description: string;
}

export const miningLasers: MiningLaser[] = [
  {
    name: "Arbor MH1",
    size: 1 as 0 | 1 | 2,
    price: 0,
    optimumRange: 60,
    maxRange: 180,
    minPower: 95,
    minPowerPct: 5,
    maxPower: 1890,
    extractPower: 1850,
    moduleSlots: 1,
    resistance: 25,
    instability: -35,
    optimalChargeRate: 0,
    optimalChargeWindow: 40,
    inertMaterials: -30,
    description: "",
  },
  {
    name: "Helix I",
    size: 1 as 0 | 1 | 2,
    price: 0,
    optimumRange: 15,
    maxRange: 45,
    minPower: 630,
    minPowerPct: 20,
    maxPower: 3150,
    extractPower: 1850,
    moduleSlots: 2,
    resistance: -30,
    instability: 0,
    optimalChargeRate: 0,
    optimalChargeWindow: -40,
    inertMaterials: -30,
    description: "",
  },
  {
    name: "Hofstede-S1",
    size: 1 as 0 | 1 | 2,
    price: 0,
    optimumRange: 45,
    maxRange: 135,
    minPower: 105,
    minPowerPct: 5,
    maxPower: 2100,
    extractPower: 1295,
    moduleSlots: 1,
    resistance: -30,
    instability: 10,
    optimalChargeRate: 20,
    optimalChargeWindow: 60,
    inertMaterials: -30,
    description: "",
  },
  {
    name: "Impact I",
    size: 1 as 0 | 1 | 2,
    price: 0,
    optimumRange: 45,
    maxRange: 135,
    minPower: 420,
    minPowerPct: 20,
    maxPower: 2100,
    extractPower: 2775,
    moduleSlots: 2,
    resistance: 10,
    instability: -10,
    optimalChargeRate: -40,
    optimalChargeWindow: 20,
    inertMaterials: -30,
    description: "",
  },
  {
    name: "Klein-S1",
    size: 1 as 0 | 1 | 2,
    price: 0,
    optimumRange: 45,
    maxRange: 135,
    minPower: 378,
    minPowerPct: 15,
    maxPower: 2520,
    extractPower: 2220,
    moduleSlots: 0,
    resistance: -45,
    instability: 35,
    optimalChargeRate: 0,
    optimalChargeWindow: 20,
    inertMaterials: -30,
    description: "",
  },
  {
    name: "Lancet MH1",
    size: 1 as 0 | 1 | 2,
    price: 0,
    optimumRange: 30,
    maxRange: 90,
    minPower: 504,
    minPowerPct: 20,
    maxPower: 2520,
    extractPower: 1850,
    moduleSlots: 1,
    resistance: 0,
    instability: -10,
    optimalChargeRate: 40,
    optimalChargeWindow: -60,
    inertMaterials: -30,
    description: "",
  },
  {
    name: "Pitman",
    size: 1 as 0 | 1 | 2,
    price: 0,
    optimumRange: 40,
    maxRange: 45,
    minPower: 630,
    minPowerPct: 20,
    maxPower: 3150,
    extractPower: 1295,
    moduleSlots: 2,
    resistance: 25,
    instability: 35,
    optimalChargeRate: -40,
    optimalChargeWindow: 40,
    inertMaterials: -40,
    description: "",
  },
  {
    name: "Arbor MH2",
    size: 2 as 0 | 1 | 2,
    price: 0,
    optimumRange: 90,
    maxRange: 270,
    minPower: 120,
    minPowerPct: 5,
    maxPower: 2400,
    extractPower: 2590,
    moduleSlots: 2,
    resistance: 25,
    instability: -35,
    optimalChargeRate: 0,
    optimalChargeWindow: 40,
    inertMaterials: -40,
    description: "",
  },
  {
    name: "Helix II",
    size: 2 as 0 | 1 | 2,
    price: 0,
    optimumRange: 30,
    maxRange: 90,
    minPower: 1224,
    minPowerPct: 30,
    maxPower: 4080,
    extractPower: 2590,
    moduleSlots: 3,
    resistance: -30,
    instability: 0,
    optimalChargeRate: 0,
    optimalChargeWindow: -40,
    inertMaterials: -40,
    description: "",
  },
  {
    name: "Hofstede-S2",
    size: 2 as 0 | 1 | 2,
    price: 0,
    optimumRange: 60,
    maxRange: 180,
    minPower: 336,
    minPowerPct: 10,
    maxPower: 3360,
    extractPower: 1295,
    moduleSlots: 2,
    resistance: -30,
    instability: 10,
    optimalChargeRate: 20,
    optimalChargeWindow: 60,
    inertMaterials: -40,
    description: "",
  },
  {
    name: "Impact II",
    size: 2 as 0 | 1 | 2,
    price: 0,
    optimumRange: 60,
    maxRange: 180,
    minPower: 1008,
    minPowerPct: 30,
    maxPower: 3360,
    extractPower: 3145,
    moduleSlots: 3,
    resistance: 10,
    instability: -10,
    optimalChargeRate: -40,
    optimalChargeWindow: 20,
    inertMaterials: -40,
    description: "",
  },
  {
    name: "Klein-S2",
    size: 2 as 0 | 1 | 2,
    price: 0,
    optimumRange: 60,
    maxRange: 180,
    minPower: 720,
    minPowerPct: 20,
    maxPower: 3600,
    extractPower: 2775,
    moduleSlots: 1,
    resistance: -45,
    instability: 35,
    optimalChargeRate: 0,
    optimalChargeWindow: 20,
    inertMaterials: -40,
    description: "",
  },
  {
    name: "Lancet MH2",
    size: 2 as 0 | 1 | 2,
    price: 0,
    optimumRange: 45,
    maxRange: 135,
    minPower: 1080,
    minPowerPct: 30,
    maxPower: 3600,
    extractPower: 2590,
    moduleSlots: 2,
    resistance: 0,
    instability: -10,
    optimalChargeRate: 40,
    optimalChargeWindow: -60,
    inertMaterials: -40,
    description: "",
  },
];